feat(files): modularised files
License: MIT
