+++ js/src/index.js
@@ -1,22 +1,22 @@
 'use strict'
 
-exports.object = require('./object')
-exports.files = require('./files')
-exports.filesMFS = require('./files-mfs')
+exports.block = require('./block')
+exports.bootstrap = require('./bootstrap')
 exports.config = require('./config')
-exports.pin = require('./pin')
+exports.dag = require('./dag')
+exports.dht = require('./dht')
+exports.files = require('./files')
+exports.key = require('./key')
+exports.ls = require('./ls')
 exports.generic = require('./miscellaneous')
 exports.miscellaneous = require('./miscellaneous')
-exports.swarm = require('./swarm')
-exports.block = require('./block')
-exports.dht = require('./dht')
-exports.dag = require('./dag')
+exports.object = require('./object')
+exports.pin = require('./pin')
 exports.ping = require('./ping')
 exports.pubsub = require('./pubsub')
-exports.key = require('./key')
-exports.stats = require('./stats')
 exports.repo = require('./repo')
-exports.bootstrap = require('./bootstrap')
+exports.stats = require('./stats')
+exports.swarm = require('./swarm')
 exports.types = require('./types')
 exports.util = require('./util')
 exports.bitswap = require('./bitswap')
